leave variables with more than 60% missing values
When there are too many missing values in a dataset, we remove such variables because they do not provide useful information. To achieve this, we can set a threshold level and remove variables with more missing values than this threshold. In our case, we leave variables with more than 60% missing values